Divide test case NodeArrow
